server/requirement.txt
# Web Framework & Server
fastapi==0.129.0
uvicorn==0.40.0
starlette==0.52.1

# AI & Environment
openai==2.21.0
python-dotenv==1.2.1

# Database & Migrations
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1

# Data Validation
pydantic==2.12.5
pydantic-settings==2.7.1
pydantic_core==2.41.5
annotated-types==0.7.0

# HTTP Clients & Utilities
httpx==0.28.1
httpcore==1.0.9
anyio==4.12.1
click==8.3.1
colorama==0.4.6
distro==1.9.0
h11==0.16.0
idna==3.11
sniffio==1.3.1
tqdm==4.67.3
typing_extensions==4.15.0
certifi==2026.1.4